move out all "item ui" to its own module
