+++ crates/re_viewer/src/ui/view_spatial/ui.rs
@@ -13,6 +13,7 @@ use crate::{
     ui::{
         data_blueprint::DataBlueprintTree,
         data_ui::{self, DataUi},
+        item_ui,
         space_view::ScreenshotMode,
         view_spatial::UiLabelTarget,
     },
@@ -246,7 +247,7 @@ impl ViewSpatialState {
                 .on_hover_text("The origin is at the origin of this Entity. All transforms are relative to it");
             // Specify space view id only if this is actually part of the space view itself.
             // (otherwise we get a somewhat broken link)
-            ctx.entity_path_button(
+            item_ui::entity_path_button(ctx,
                 ui,
                 data_blueprint
                     .contains_entity(space_path)
@@ -848,7 +849,7 @@ pub fn picking(
         } else {
             // Hover ui for everything else
             response.on_hover_ui_at_pointer(|ui| {
-                ctx.instance_path_button(ui, Some(space_view_id), &instance_path);
+                item_ui::instance_path_button(ctx, ui, Some(space_view_id), &instance_path);
                 instance_path.data_ui(
                     ctx,
                     ui,
